Migrations

+++ database/migrations/2022_08_01_202355_create_professionals_table.php
@@ -13,29 +13,31 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('professionals', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('professional')) {
+            Schema::create('professionals', function (Blueprint $table) {
+                $table->id();
 
-            $table->string('first_name');
-            $table->string('last_name');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->rememberToken();
+                $table->string('first_name');
+                $table->string('last_name');
+                $table->string('email')->unique();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->string('password');
+                $table->rememberToken();
 
-            $table->string('api_token')->nullable();
-            $table->string('profile_photo_path')->default('https://www.nicepng.com/png/detail/933-9332131_profile-picture-default-png.png')->nullable();
-            $table->string('phone_number')->nullable();
-            $table->string('address_line_one')->nullable();
-            $table->string('address_line_two')->nullable();
-            $table->string('zip_code')->nullable();
-            $table->string('country')->nullable();
+                $table->string('api_token')->nullable();
+                $table->string('profile_photo_path')->default('https://www.nicepng.com/png/detail/933-9332131_profile-picture-default-png.png')->nullable();
+                $table->string('phone_number')->nullable();
+                $table->string('address_line_one')->nullable();
+                $table->string('address_line_two')->nullable();
+                $table->string('zip_code')->nullable();
+                $table->string('country')->nullable();
 
-            $table->unsignedBigInteger('company_id');
-            $table->foreign('company_id')->references('id')->on('companies')->onUpdate('cascade')->onDelete('cascade');
+                $table->unsignedBigInteger('company_id');
+                $table->foreign('company_id')->references('id')->on('companies')->onUpdate('cascade')->onDelete('cascade');
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
