+++ database/migrations/2022_08_01_223919_create_establishments_table.php
@@ -13,24 +13,26 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('establishments', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('establishments')) {
+            Schema::create('establishments', function (Blueprint $table) {
+                $table->id();
 
-            $table->unsignedBigInteger('company_id');
-            $table->foreign('company_id')->references('id')->on('companies')->onUpdate('cascade')->onDelete('cascade');
+                $table->unsignedBigInteger('company_id');
+                $table->foreign('company_id')->references('id')->on('companies')->onUpdate('cascade')->onDelete('cascade');
 
-            $table->string('name');
-            $table->string('slogan')->nullable();
-            $table->string('type')->nullable();
-            $table->string('address')->nullable();
-            $table->string('city')->nullable();
-            $table->string('zip_code')->nullable();
-            $table->string('longitude')->nullable();
-            $table->string('latitude')->nullable();
-            $table->string('description')->nullable();
+                $table->string('name');
+                $table->string('slogan')->nullable();
+                $table->string('type')->nullable();
+                $table->string('address')->nullable();
+                $table->string('city')->nullable();
+                $table->string('zip_code')->nullable();
+                $table->string('longitude')->nullable();
+                $table->string('latitude')->nullable();
+                $table->string('description')->nullable();
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -40,6 +42,6 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('establishments');
+        // Schema::dropIfExists('establishments');
     }
 };
